+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 * nostr: add `Coordinate::verify` method ([Yuki Kishimoto])
 * nostr: add `TagStandard::Client` variant ([Yuki Kishimoto])
 * nostr: add `EventBuilder::dedup_tags` method ([Yuki Kishimoto])
+* nostr: impl `FromIterator<Tag>` for `Tags` ([Yuki Kishimoto])
 * pool: event verification cache ([Yuki Kishimoto])
 * ffi: add Mac Catalyst support in Swift package ([Yuki Kishimoto])